backend/requirements.txt
python-dotenv
fastapi
uvicorn
pydantic
httpx
numpy
qdrant-client==1.4.0
huggingface-hub==0.16.4
transformers
groq
langchain[google-genai]
langchain[groq]
langchain-google-genai
langchain-core
langgraph 
langchain-community
#llama-cpp-python==0.1.77